services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.20.0",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "fluent-logger>=0.10.0",
    "greenlet>=3.3.0",
    "httpx>=0.27.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",
    "prometheus-client>=0.20.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "structlog>=24.4.0",
    "temporalio>=1.8.0",
    "uvicorn>=0.40.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
